pyproject.toml
[tool.poetry]
name = "chatter"
version = "0.1.0"
description = ""
authors = ["Chris Birch <datapointchris@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.38.0"
streamlit = "^1.37.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ipywidgets = "^8.1.3"
black = "^24.8.0"
flake8 = "^7.1.0"
pytest = "^8.3.2"
pre-commit = "^3.8.0"
pytest-cov = "^5.0.0"
bandit = "^1.7.9"
mypy = "^1.11.1"
isort = "^5.13.2"

[tool.mypy]
pretty = true
ignore_missing_imports = true
warn_return_any = false

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
